backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "pandas>=2.2.3",
    "psycopg[binary]>=3.2.6",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
    "yfinance>=0.2.56",
]
